Develop: Fix ESLint errors.
Disable react/no-array-index-key in Board
as it's not applicable in this instance.
Because the only information we have about each cell is its index
and because the cells won't change index.

Disable security/detect-object-injection, as it's safe in this case
because it's just a string that doesn't execute anywhere else.